requirements.txt
mysql_connector_repackaged==0.3.1
pandas==2.3.3
numpy==2.1.1
plotly==5.24.1
reportlab==4.4.4
streamlit==1.37.1
streamlit_option_menu==0.4.0
requests==2.32.3

